Moved Integer and Long functionality into classes extending AbstractTestUtil
