+++ src/main/java/com/threeleaf/test/random/TestNumber.java
@@ -1,5 +1,6 @@
 package com.threeleaf.test.random;
 
+import com.threeleaf.test.random.util.RandomIntegerUtil;
 import lombok.experimental.UtilityClass;
 
 import java.math.BigDecimal;
@@ -22,7 +23,7 @@ public class TestNumber
     public static final BigDecimal BIG_DECIMAL_RANDOM = randomBigDecimal();
 
     /** A random {@link BigDecimal} from 0-100. */
-    public static final BigDecimal BIG_DECIMAL_SMALL = valueOf(randomBetween(1, 10000) / 100D).setScale(2, ROUND_HALF_UP);
+    public static final BigDecimal BIG_DECIMAL_SMALL = valueOf(RandomIntegerUtil.randomBetween(1, 10000) / 100D).setScale(2, ROUND_HALF_UP);
 
     /** A random {@link BigInteger}. */
     public static final BigInteger BIG_INTEGER = randomBigInteger();
@@ -31,7 +32,7 @@ public class TestNumber
     public static final BigInteger BIG_INTEGER_RANDOM = randomBigInteger();
 
     /** A random {@link BigInteger} from 0-100. */
-    public static final BigInteger BIG_INTEGER_SMALL = BigInteger.valueOf(randomBetween(0, 100));
+    public static final BigInteger BIG_INTEGER_SMALL = BigInteger.valueOf(RandomIntegerUtil.randomBetween(0, 100));
 
     /** A random Byte. */
     public final static Byte BYTE = randomByte();
@@ -352,7 +353,7 @@ public class TestNumber
     public final static Byte BYTE_RANDOM = randomByte();
 
     /** A small Byte from 0-100. */
-    public final static Byte BYTE_SMALL = (byte) randomBetween(0, 100);
+    public final static Byte BYTE_SMALL = (byte) RandomIntegerUtil.randomBetween(0, 100);
 
     /** A random Double. */
     public final static Double DOUBLE = randomDouble();
@@ -694,7 +695,7 @@ public class TestNumber
     public final static Double DOUBLE_RANDOM = randomDouble();
 
     /** A small Double between 0-100. */
-    public final static Double DOUBLE_SMALL = randomBetween(0, 10000) / 100D;
+    public final static Double DOUBLE_SMALL = RandomIntegerUtil.randomBetween(0, 10000) / 100D;
 
     /** A random Float. */
     public final static Float FLOAT = randomFloat();
@@ -1033,691 +1034,7 @@ public class TestNumber
     public final static Float FLOAT_RANDOM = randomFloat();
 
     /** A small Float between 0-100. */
-    public final static Float FLOAT_SMALL = randomBetween(0, 10000) / 100F;
-
-    /** A random Integer. */
-    public final static Integer INTEGER = randomInt();
-
-    /** The Integer 00. */
-    public final static Integer INTEGER_00 = 0;
-
-    /** The Integer 01. */
-    public final static Integer INTEGER_01 = 1;
-
-    /** The Integer 02. */
-    public final static Integer INTEGER_02 = 2;
-
-    /** The Integer 03. */
-    public final static Integer INTEGER_03 = 3;
-
-    /** The Integer 04. */
-    public final static Integer INTEGER_04 = 4;
-
-    /** The Integer 05. */
-    public final static Integer INTEGER_05 = 5;
-
-    /** The Integer 06. */
-    public final static Integer INTEGER_06 = 6;
-
-    /** The Integer 07. */
-    public final static Integer INTEGER_07 = 7;
-
-    /** The Integer 08. */
-    public final static Integer INTEGER_08 = 8;
-
-    /** The Integer 09. */
-    public final static Integer INTEGER_09 = 9;
-
-    /** The Integer 10. */
-    public final static Integer INTEGER_10 = 10;
-
-    /** The Integer 100. */
-    public final static Integer INTEGER_100 = 100;
-
-    /** The Integer 1000. */
-    public final static Integer INTEGER_1000 = 1000;
-
-    /** The Integer 10000. */
-    public final static Integer INTEGER_10000 = 10000;
-
-    /** The Integer 10001. */
-    public final static Integer INTEGER_10001 = 10001;
-
-    /** The Integer 1001. */
-    public final static Integer INTEGER_1001 = 1001;
-
-    /** The Integer 101. */
-    public final static Integer INTEGER_101 = 101;
-
-    /** The Integer 11. */
-    public final static Integer INTEGER_11 = 11;
-
-    /** The Integer 12. */
-    public final static Integer INTEGER_12 = 12;
-
-    /** The Integer 13. */
-    public final static Integer INTEGER_13 = 13;
-
-    /** The Integer 14. */
-    public final static Integer INTEGER_14 = 14;
-
-    /** The Integer 15. */
-    public final static Integer INTEGER_15 = 15;
-
-    /** The Integer 16. */
-    public final static Integer INTEGER_16 = 16;
-
-    /** The Integer 17. */
-    public final static Integer INTEGER_17 = 17;
-
-    /** The Integer 18. */
-    public final static Integer INTEGER_18 = 18;
-
-    /** The Integer 19. */
-    public final static Integer INTEGER_19 = 19;
-
-    /** The Integer 20. */
-    public final static Integer INTEGER_20 = 20;
-
-    /** The Integer 21. */
-    public final static Integer INTEGER_21 = 21;
-
-    /** The Integer 22. */
-    public final static Integer INTEGER_22 = 22;
-
-    /** The Integer 23. */
-    public final static Integer INTEGER_23 = 23;
-
-    /** The Integer 24. */
-    public final static Integer INTEGER_24 = 24;
-
-    /** The Integer 25. */
-    public final static Integer INTEGER_25 = 25;
-
-    /** The Integer 26. */
-    public final static Integer INTEGER_26 = 26;
-
-    /** The Integer 27. */
-    public final static Integer INTEGER_27 = 27;
-
-    /** The Integer 28. */
-    public final static Integer INTEGER_28 = 28;
-
-    /** The Integer 29. */
-    public final static Integer INTEGER_29 = 29;
-
-    /** The Integer 30. */
-    public final static Integer INTEGER_30 = 30;
-
-    /** The Integer 31. */
-    public final static Integer INTEGER_31 = 31;
-
-    /** The Integer 32. */
-    public final static Integer INTEGER_32 = 32;
-
-    /** The Integer 33. */
-    public final static Integer INTEGER_33 = 33;
-
-    /** The Integer 34. */
-    public final static Integer INTEGER_34 = 34;
-
-    /** The Integer 35. */
-    public final static Integer INTEGER_35 = 35;
-
-    /** The Integer 36. */
-    public final static Integer INTEGER_36 = 36;
-
-    /** The Integer 37. */
-    public final static Integer INTEGER_37 = 37;
-
-    /** The Integer 38. */
-    public final static Integer INTEGER_38 = 38;
-
-    /** The Integer 39. */
-    public final static Integer INTEGER_39 = 39;
-
-    /** The Integer 40. */
-    public final static Integer INTEGER_40 = 40;
-
-    /** The Integer 41. */
-    public final static Integer INTEGER_41 = 41;
-
-    /** The Integer 42. */
-    public final static Integer INTEGER_42 = 42;
-
-    /** The Integer 43. */
-    public final static Integer INTEGER_43 = 43;
-
-    /** The Integer 44. */
-    public final static Integer INTEGER_44 = 44;
-
-    /** The Integer 45. */
-    public final static Integer INTEGER_45 = 45;
-
-    /** The Integer 46. */
-    public final static Integer INTEGER_46 = 46;
-
-    /** The Integer 47. */
-    public final static Integer INTEGER_47 = 47;
-
-    /** The Integer 48. */
-    public final static Integer INTEGER_48 = 48;
-
-    /** The Integer 49. */
-    public final static Integer INTEGER_49 = 49;
-
-    /** The Integer 50. */
-    public final static Integer INTEGER_50 = 50;
-
-    /** The Integer 51. */
-    public final static Integer INTEGER_51 = 51;
-
-    /** The Integer 52. */
-    public final static Integer INTEGER_52 = 52;
-
-    /** The Integer 53. */
-    public final static Integer INTEGER_53 = 53;
-
-    /** The Integer 54. */
-    public final static Integer INTEGER_54 = 54;
-
-    /** The Integer 55. */
-    public final static Integer INTEGER_55 = 55;
-
-    /** The Integer 56. */
-    public final static Integer INTEGER_56 = 56;
-
-    /** The Integer 57. */
-    public final static Integer INTEGER_57 = 57;
-
-    /** The Integer 58. */
-    public final static Integer INTEGER_58 = 58;
-
-    /** The Integer 59. */
-    public final static Integer INTEGER_59 = 59;
-
-    /** The Integer 60. */
-    public final static Integer INTEGER_60 = 60;
-
-    /** The Integer 61. */
-    public final static Integer INTEGER_61 = 61;
-
-    /** The Integer 62. */
-    public final static Integer INTEGER_62 = 62;
-
-    /** The Integer 63. */
-    public final static Integer INTEGER_63 = 63;
-
-    /** The Integer 64. */
-    public final static Integer INTEGER_64 = 64;
-
-    /** The Integer 65. */
-    public final static Integer INTEGER_65 = 65;
-
-    /** The Integer 66. */
-    public final static Integer INTEGER_66 = 66;
-
-    /** The Integer 67. */
-    public final static Integer INTEGER_67 = 67;
-
-    /** The Integer 68. */
-    public final static Integer INTEGER_68 = 68;
-
-    /** The Integer 69. */
-    public final static Integer INTEGER_69 = 69;
-
-    /** The Integer 70. */
-    public final static Integer INTEGER_70 = 70;
-
-    /** The Integer 71. */
-    public final static Integer INTEGER_71 = 71;
-
-    /** The Integer 72. */
-    public final static Integer INTEGER_72 = 72;
-
-    /** The Integer 73. */
-    public final static Integer INTEGER_73 = 73;
-
-    /** The Integer 74. */
-    public final static Integer INTEGER_74 = 74;
-
-    /** The Integer 75. */
-    public final static Integer INTEGER_75 = 75;
-
-    /** The Integer 76. */
-    public final static Integer INTEGER_76 = 76;
-
-    /** The Integer 77. */
-    public final static Integer INTEGER_77 = 77;
-
-    /** The Integer 78. */
-    public final static Integer INTEGER_78 = 78;
-
-    /** The Integer 79. */
-    public final static Integer INTEGER_79 = 79;
-
-    /** The Integer 80. */
-    public final static Integer INTEGER_80 = 80;
-
-    /** The Integer 81. */
-    public final static Integer INTEGER_81 = 81;
-
-    /** The Integer 82. */
-    public final static Integer INTEGER_82 = 82;
-
-    /** The Integer 83. */
-    public final static Integer INTEGER_83 = 83;
-
-    /** The Integer 84. */
-    public final static Integer INTEGER_84 = 84;
-
-    /** The Integer 85. */
-    public final static Integer INTEGER_85 = 85;
-
-    /** The Integer 86. */
-    public final static Integer INTEGER_86 = 86;
-
-    /** The Integer 87. */
-    public final static Integer INTEGER_87 = 87;
-
-    /** The Integer 88. */
-    public final static Integer INTEGER_88 = 88;
-
-    /** The Integer 89. */
-    public final static Integer INTEGER_89 = 89;
-
-    /** The Integer 90. */
-    public final static Integer INTEGER_90 = 90;
-
-    /** The Integer 91. */
-    public final static Integer INTEGER_91 = 91;
-
-    /** The Integer 92. */
-    public final static Integer INTEGER_92 = 92;
-
-    /** The Integer 93. */
-    public final static Integer INTEGER_93 = 93;
-
-    /** The Integer 94. */
-    public final static Integer INTEGER_94 = 94;
-
-    /** The Integer 95. */
-    public final static Integer INTEGER_95 = 95;
-
-    /** The Integer 96. */
-    public final static Integer INTEGER_96 = 96;
-
-    /** The Integer 97. */
-    public final static Integer INTEGER_97 = 97;
-
-    /** The Integer 98. */
-    public final static Integer INTEGER_98 = 98;
-
-    /** The Integer 99. */
-    public final static Integer INTEGER_99 = 99;
-
-    /** The Integer 999. */
-    public final static Integer INTEGER_999 = 999;
-
-    /** The Integer 9999. */
-    public final static Integer INTEGER_9999 = 9999;
-
-    /** A negative Integer. */
-    public final static Integer INTEGER_NEGATIVE = -Math.abs(randomInt());
-
-    /** A positive Integer. */
-    public final static Integer INTEGER_POSITIVE = Math.abs(randomInt());
-
-    /** A random Integer. */
-    public final static Integer INTEGER_RANDOM = randomInt();
-
-    /** A small Integer between 0-100. */
-    public final static Integer INTEGER_SMALL = randomBetween(0, 100);
-
-    /** A random Long. */
-    public final static Long LONG = randomLong();
-
-    /** The Long 00. */
-    public final static Long LONG_00 = 0L;
-
-    /** The Long 01. */
-    public final static Long LONG_01 = 1L;
-
-    /** The Long 02. */
-    public final static Long LONG_02 = 2L;
-
-    /** The Long 03. */
-    public final static Long LONG_03 = 3L;
-
-    /** The Long 04. */
-    public final static Long LONG_04 = 4L;
-
-    /** The Long 05. */
-    public final static Long LONG_05 = 5L;
-
-    /** The Long 06. */
-    public final static Long LONG_06 = 6L;
-
-    /** The Long 07. */
-    public final static Long LONG_07 = 7L;
-
-    /** The Long 08. */
-    public final static Long LONG_08 = 8L;
-
-    /** The Long 09. */
-    public final static Long LONG_09 = 9L;
-
-    /** The Long 10. */
-    public final static Long LONG_10 = 10L;
-
-    /** The Long 100. */
-    public final static Long LONG_100 = 100L;
-
-    /** The Long 1000. */
-    public final static Long LONG_1000 = 1000L;
-
-    /** The Long 10000. */
-    public final static Long LONG_10000 = 10000L;
-
-    /** The Long 10001. */
-    public final static Long LONG_10001 = 10001L;
-
-    /** The Long 1001. */
-    public final static Long LONG_1001 = 1001L;
-
-    /** The Long 101. */
-    public final static Long LONG_101 = 101L;
-
-    /** The Long 11. */
-    public final static Long LONG_11 = 11L;
-
-    /** The Long 12. */
-    public final static Long LONG_12 = 12L;
-
-    /** The Long 13. */
-    public final static Long LONG_13 = 13L;
-
-    /** The Long 14. */
-    public final static Long LONG_14 = 14L;
-
-    /** The Long 15. */
-    public final static Long LONG_15 = 15L;
-
-    /** The Long 16. */
-    public final static Long LONG_16 = 16L;
-
-    /** The Long 17. */
-    public final static Long LONG_17 = 17L;
-
-    /** The Long 18. */
-    public final static Long LONG_18 = 18L;
-
-    /** The Long 19. */
-    public final static Long LONG_19 = 19L;
-
-    /** The Long 20. */
-    public final static Long LONG_20 = 20L;
-
-    /** The Long 21. */
-    public final static Long LONG_21 = 21L;
-
-    /** The Long 22. */
-    public final static Long LONG_22 = 22L;
-
-    /** The Long 23. */
-    public final static Long LONG_23 = 23L;
-
-    /** The Long 24. */
-    public final static Long LONG_24 = 24L;
-
-    /** The Long 25. */
-    public final static Long LONG_25 = 25L;
-
-    /** The Long 26. */
-    public final static Long LONG_26 = 26L;
-
-    /** The Long 27. */
-    public final static Long LONG_27 = 27L;
-
-    /** The Long 28. */
-    public final static Long LONG_28 = 28L;
-
-    /** The Long 29. */
-    public final static Long LONG_29 = 29L;
-
-    /** The Long 30. */
-    public final static Long LONG_30 = 30L;
-
-    /** The Long 31. */
-    public final static Long LONG_31 = 31L;
-
-    /** The Long 32. */
-    public final static Long LONG_32 = 32L;
-
-    /** The Long 33. */
-    public final static Long LONG_33 = 33L;
-
-    /** The Long 34. */
-    public final static Long LONG_34 = 34L;
-
-    /** The Long 35. */
-    public final static Long LONG_35 = 35L;
-
-    /** The Long 36. */
-    public final static Long LONG_36 = 36L;
-
-    /** The Long 37. */
-    public final static Long LONG_37 = 37L;
-
-    /** The Long 38. */
-    public final static Long LONG_38 = 38L;
-
-    /** The Long 39. */
-    public final static Long LONG_39 = 39L;
-
-    /** The Long 40. */
-    public final static Long LONG_40 = 40L;
-
-    /** The Long 41. */
-    public final static Long LONG_41 = 41L;
-
-    /** The Long 42. */
-    public final static Long LONG_42 = 42L;
-
-    /** The Long 43. */
-    public final static Long LONG_43 = 43L;
-
-    /** The Long 44. */
-    public final static Long LONG_44 = 44L;
-
-    /** The Long 45. */
-    public final static Long LONG_45 = 45L;
-
-    /** The Long 46. */
-    public final static Long LONG_46 = 46L;
-
-    /** The Long 47. */
-    public final static Long LONG_47 = 47L;
-
-    /** The Long 48. */
-    public final static Long LONG_48 = 48L;
-
-    /** The Long 49. */
-    public final static Long LONG_49 = 49L;
-
-    /** The Long 50. */
-    public final static Long LONG_50 = 50L;
-
-    /** The Long 51. */
-    public final static Long LONG_51 = 51L;
-
-    /** The Long 52. */
-    public final static Long LONG_52 = 52L;
-
-    /** The Long 53. */
-    public final static Long LONG_53 = 53L;
-
-    /** The Long 54. */
-    public final static Long LONG_54 = 54L;
-
-    /** The Long 55. */
-    public final static Long LONG_55 = 55L;
-
-    /** The Long 56. */
-    public final static Long LONG_56 = 56L;
-
-    /** The Long 57. */
-    public final static Long LONG_57 = 57L;
-
-    /** The Long 58. */
-    public final static Long LONG_58 = 58L;
-
-    /** The Long 59. */
-    public final static Long LONG_59 = 59L;
-
-    /** The Long 60. */
-    public final static Long LONG_60 = 60L;
-
-    /** The Long 61. */
-    public final static Long LONG_61 = 61L;
-
-    /** The Long 62. */
-    public final static Long LONG_62 = 62L;
-
-    /** The Long 63. */
-    public final static Long LONG_63 = 63L;
-
-    /** The Long 64. */
-    public final static Long LONG_64 = 64L;
-
-    /** The Long 65. */
-    public final static Long LONG_65 = 65L;
-
-    /** The Long 66. */
-    public final static Long LONG_66 = 66L;
-
-    /** The Long 67. */
-    public final static Long LONG_67 = 67L;
-
-    /** The Long 68. */
-    public final static Long LONG_68 = 68L;
-
-    /** The Long 69. */
-    public final static Long LONG_69 = 69L;
-
-    /** The Long 70. */
-    public final static Long LONG_70 = 70L;
-
-    /** The Long 71. */
-    public final static Long LONG_71 = 71L;
-
-    /** The Long 72. */
-    public final static Long LONG_72 = 72L;
-
-    /** The Long 73. */
-    public final static Long LONG_73 = 73L;
-
-    /** The Long 74. */
-    public final static Long LONG_74 = 74L;
-
-    /** The Long 75. */
-    public final static Long LONG_75 = 75L;
-
-    /** The Long 76. */
-    public final static Long LONG_76 = 76L;
-
-    /** The Long 77. */
-    public final static Long LONG_77 = 77L;
-
-    /** The Long 78. */
-    public final static Long LONG_78 = 78L;
-
-    /** The Long 79. */
-    public final static Long LONG_79 = 79L;
-
-    /** The Long 80. */
-    public final static Long LONG_80 = 80L;
-
-    /** The Long 81. */
-    public final static Long LONG_81 = 81L;
-
-    /** The Long 82. */
-    public final static Long LONG_82 = 82L;
-
-    /** The Long 83. */
-    public final static Long LONG_83 = 83L;
-
-    /** The Long 84. */
-    public final static Long LONG_84 = 84L;
-
-    /** The Long 85. */
-    public final static Long LONG_85 = 85L;
-
-    /** The Long 86. */
-    public final static Long LONG_86 = 86L;
-
-    /** The Long 87. */
-    public final static Long LONG_87 = 87L;
-
-    /** The Long 88. */
-    public final static Long LONG_88 = 88L;
-
-    /** The Long 89. */
-    public final static Long LONG_89 = 89L;
-
-    /** The Long 90. */
-    public final static Long LONG_90 = 90L;
-
-    /** The Long 91. */
-    public final static Long LONG_91 = 91L;
-
-    /** The Long 92. */
-    public final static Long LONG_92 = 92L;
-
-    /** The Long 93. */
-    public final static Long LONG_93 = 93L;
-
-    /** The Long 94. */
-    public final static Long LONG_94 = 94L;
-
-    /** The Long 95. */
-    public final static Long LONG_95 = 95L;
-
-    /** The Long 96. */
-    public final static Long LONG_96 = 96L;
-
-    /** The Long 97. */
-    public final static Long LONG_97 = 97L;
-
-    /** The Long 98. */
-    public final static Long LONG_98 = 98L;
-
-    /** The Long 99. */
-    public final static Long LONG_99 = 99L;
-
-    /** The Long 999. */
-    public final static Long LONG_999 = 999L;
-
-    /** The Long 9999. */
-    public final static Long LONG_9999 = 9999L;
-
-    /** A negative Long. */
-    public final static Long LONG_NEGATIVE = -Math.abs(randomLong());
-
-    /** A positive Long. */
-    public final static Long LONG_POSITIVE = Math.abs(randomLong());
-
-    /** A random Long. */
-    public final static Long LONG_RANDOM = randomLong();
-
-    /** A small Long between 0-100. */
-    public final static Long LONG_SMALL = (long) randomBetween(0, 100);
-
-    /** Maximum digit value. */
-    public static final int MAX_DIGIT = 9;
-
-    /** Minimum digit value. */
-    public static final int MIN_DIGIT = 0;
+    public final static Float FLOAT_SMALL = RandomIntegerUtil.randomBetween(0, 10000) / 100F;
 
     /** A random Short. */
     public final static Short SHORT = randomShort();
@@ -2056,5 +1373,6 @@ public class TestNumber
     public final static Short SHORT_RANDOM = randomShort();
 
     /** A small Short between 0-100. */
-    public final static Short SHORT_SMALL = (short) randomBetween(0, 100);
+    public final static Short SHORT_SMALL = (short) RandomIntegerUtil.randomBetween(0, 100);
+
 }
