+++ src/main/java/com/threeleaf/test/random/TestPrimitive.java
@@ -1,5 +1,6 @@
 package com.threeleaf.test.random;
 
+import com.threeleaf.test.random.util.RandomIntegerUtil;
 import lombok.experimental.UtilityClass;
 
 import static com.threeleaf.test.random.util.TestNumberUtil.*;
@@ -326,7 +327,7 @@ public class TestPrimitive
     public final static byte B_RANDOM = randomByte();
 
     /** A small primitive byte from 0-100. */
-    public final static byte B_SMALL = (byte) randomBetween(0, 100);
+    public final static byte B_SMALL = (byte) RandomIntegerUtil.randomBetween(0, 100);
 
     /** The primitive double 00. */
     public final static double D_00 = 0;
@@ -662,7 +663,7 @@ public class TestPrimitive
     public final static double D_RANDOM = randomDouble();
 
     /** A small primitive double between 0-100. */
-    public final static double D_SMALL = randomBetween(0, 10000) / 100D;
+    public final static double D_SMALL = RandomIntegerUtil.randomBetween(0, 10000) / 100D;
 
     /** The primitive float 00. */
     public final static float F_00 = 0;
@@ -998,679 +999,7 @@ public class TestPrimitive
     public final static float F_RANDOM = randomFloat();
 
     /** A small primitive float between 0-100. */
-    public final static float F_SMALL = randomBetween(0, 10000) / 100F;
-
-    /** The primitive int 00. */
-    public final static int INT_00 = 0;
-
-    /** The primitive int 01. */
-    public final static int INT_01 = 1;
-
-    /** The primitive int 02. */
-    public final static int INT_02 = 2;
-
-    /** The primitive int 03. */
-    public final static int INT_03 = 3;
-
-    /** The primitive int 04. */
-    public final static int INT_04 = 4;
-
-    /** The primitive int 05. */
-    public final static int INT_05 = 5;
-
-    /** The primitive int 06. */
-    public final static int INT_06 = 6;
-
-    /** The primitive int 07. */
-    public final static int INT_07 = 7;
-
-    /** The primitive int 08. */
-    public final static int INT_08 = 8;
-
-    /** The primitive int 09. */
-    public final static int INT_09 = 9;
-
-    /** The primitive int 10. */
-    public final static int INT_10 = 10;
-
-    /** The primitive int 100. */
-    public final static int INT_100 = 100;
-
-    /** The primitive int 1000. */
-    public final static int INT_1000 = 1000;
-
-    /** The primitive int 10000. */
-    public final static int INT_10000 = 10000;
-
-    /** The primitive int 10001. */
-    public final static int INT_10001 = 10001;
-
-    /** The primitive int 1001. */
-    public final static int INT_1001 = 1001;
-
-    /** The primitive int 101. */
-    public final static int INT_101 = 101;
-
-    /** The primitive int 11. */
-    public final static int INT_11 = 11;
-
-    /** The primitive int 12. */
-    public final static int INT_12 = 12;
-
-    /** The primitive int 13. */
-    public final static int INT_13 = 13;
-
-    /** The primitive int 14. */
-    public final static int INT_14 = 14;
-
-    /** The primitive int 15. */
-    public final static int INT_15 = 15;
-
-    /** The primitive int 16. */
-    public final static int INT_16 = 16;
-
-    /** The primitive int 17. */
-    public final static int INT_17 = 17;
-
-    /** The primitive int 18. */
-    public final static int INT_18 = 18;
-
-    /** The primitive int 19. */
-    public final static int INT_19 = 19;
-
-    /** The primitive int 20. */
-    public final static int INT_20 = 20;
-
-    /** The primitive int 21. */
-    public final static int INT_21 = 21;
-
-    /** The primitive int 22. */
-    public final static int INT_22 = 22;
-
-    /** The primitive int 23. */
-    public final static int INT_23 = 23;
-
-    /** The primitive int 24. */
-    public final static int INT_24 = 24;
-
-    /** The primitive int 25. */
-    public final static int INT_25 = 25;
-
-    /** The primitive int 26. */
-    public final static int INT_26 = 26;
-
-    /** The primitive int 27. */
-    public final static int INT_27 = 27;
-
-    /** The primitive int 28. */
-    public final static int INT_28 = 28;
-
-    /** The primitive int 29. */
-    public final static int INT_29 = 29;
-
-    /** The primitive int 30. */
-    public final static int INT_30 = 30;
-
-    /** The primitive int 31. */
-    public final static int INT_31 = 31;
-
-    /** The primitive int 32. */
-    public final static int INT_32 = 32;
-
-    /** The primitive int 33. */
-    public final static int INT_33 = 33;
-
-    /** The primitive int 34. */
-    public final static int INT_34 = 34;
-
-    /** The primitive int 35. */
-    public final static int INT_35 = 35;
-
-    /** The primitive int 36. */
-    public final static int INT_36 = 36;
-
-    /** The primitive int 37. */
-    public final static int INT_37 = 37;
-
-    /** The primitive int 38. */
-    public final static int INT_38 = 38;
-
-    /** The primitive int 39. */
-    public final static int INT_39 = 39;
-
-    /** The primitive int 40. */
-    public final static int INT_40 = 40;
-
-    /** The primitive int 41. */
-    public final static int INT_41 = 41;
-
-    /** The primitive int 42. */
-    public final static int INT_42 = 42;
-
-    /** The primitive int 43. */
-    public final static int INT_43 = 43;
-
-    /** The primitive int 44. */
-    public final static int INT_44 = 44;
-
-    /** The primitive int 45. */
-    public final static int INT_45 = 45;
-
-    /** The primitive int 46. */
-    public final static int INT_46 = 46;
-
-    /** The primitive int 47. */
-    public final static int INT_47 = 47;
-
-    /** The primitive int 48. */
-    public final static int INT_48 = 48;
-
-    /** The primitive int 49. */
-    public final static int INT_49 = 49;
-
-    /** The primitive int 50. */
-    public final static int INT_50 = 50;
-
-    /** The primitive int 51. */
-    public final static int INT_51 = 51;
-
-    /** The primitive int 52. */
-    public final static int INT_52 = 52;
-
-    /** The primitive int 53. */
-    public final static int INT_53 = 53;
-
-    /** The primitive int 54. */
-    public final static int INT_54 = 54;
-
-    /** The primitive int 55. */
-    public final static int INT_55 = 55;
-
-    /** The primitive int 56. */
-    public final static int INT_56 = 56;
-
-    /** The primitive int 57. */
-    public final static int INT_57 = 57;
-
-    /** The primitive int 58. */
-    public final static int INT_58 = 58;
-
-    /** The primitive int 59. */
-    public final static int INT_59 = 59;
-
-    /** The primitive int 60. */
-    public final static int INT_60 = 60;
-
-    /** The primitive int 61. */
-    public final static int INT_61 = 61;
-
-    /** The primitive int 62. */
-    public final static int INT_62 = 62;
-
-    /** The primitive int 63. */
-    public final static int INT_63 = 63;
-
-    /** The primitive int 64. */
-    public final static int INT_64 = 64;
-
-    /** The primitive int 65. */
-    public final static int INT_65 = 65;
-
-    /** The primitive int 66. */
-    public final static int INT_66 = 66;
-
-    /** The primitive int 67. */
-    public final static int INT_67 = 67;
-
-    /** The primitive int 68. */
-    public final static int INT_68 = 68;
-
-    /** The primitive int 69. */
-    public final static int INT_69 = 69;
-
-    /** The primitive int 70. */
-    public final static int INT_70 = 70;
-
-    /** The primitive int 71. */
-    public final static int INT_71 = 71;
-
-    /** The primitive int 72. */
-    public final static int INT_72 = 72;
-
-    /** The primitive int 73. */
-    public final static int INT_73 = 73;
-
-    /** The primitive int 74. */
-    public final static int INT_74 = 74;
-
-    /** The primitive int 75. */
-    public final static int INT_75 = 75;
-
-    /** The primitive int 76. */
-    public final static int INT_76 = 76;
-
-    /** The primitive int 77. */
-    public final static int INT_77 = 77;
-
-    /** The primitive int 78. */
-    public final static int INT_78 = 78;
-
-    /** The primitive int 79. */
-    public final static int INT_79 = 79;
-
-    /** The primitive int 80. */
-    public final static int INT_80 = 80;
-
-    /** The primitive int 81. */
-    public final static int INT_81 = 81;
-
-    /** The primitive int 82. */
-    public final static int INT_82 = 82;
-
-    /** The primitive int 83. */
-    public final static int INT_83 = 83;
-
-    /** The primitive int 84. */
-    public final static int INT_84 = 84;
-
-    /** The primitive int 85. */
-    public final static int INT_85 = 85;
-
-    /** The primitive int 86. */
-    public final static int INT_86 = 86;
-
-    /** The primitive int 87. */
-    public final static int INT_87 = 87;
-
-    /** The primitive int 88. */
-    public final static int INT_88 = 88;
-
-    /** The primitive int 89. */
-    public final static int INT_89 = 89;
-
-    /** The primitive int 90. */
-    public final static int INT_90 = 90;
-
-    /** The primitive int 91. */
-    public final static int INT_91 = 91;
-
-    /** The primitive int 92. */
-    public final static int INT_92 = 92;
-
-    /** The primitive int 93. */
-    public final static int INT_93 = 93;
-
-    /** The primitive int 94. */
-    public final static int INT_94 = 94;
-
-    /** The primitive int 95. */
-    public final static int INT_95 = 95;
-
-    /** The primitive int 96. */
-    public final static int INT_96 = 96;
-
-    /** The primitive int 97. */
-    public final static int INT_97 = 97;
-
-    /** The primitive int 98. */
-    public final static int INT_98 = 98;
-
-    /** The primitive int 99. */
-    public final static int INT_99 = 99;
-
-    /** The primitive int 999. */
-    public final static int INT_999 = 999;
-
-    /** The primitive int 9999. */
-    public final static int INT_9999 = 9999;
-
-    /** A negative primitive int. */
-    public final static int INT_NEGATIVE = -Math.abs(randomInt());
-
-    /** A positive primitive int. */
-    public final static int INT_POSITIVE = Math.abs(randomInt());
-
-    /** A random primitive int. */
-    public final static int INT_RANDOM = randomInt();
-
-    /** A small primitive int between 0-100. */
-    public final static int INT_SMALL = randomBetween(0, 100);
-
-    /** The primitive long 00. */
-    public final static long L_00 = 0;
-
-    /** The primitive long 01. */
-    public final static long L_01 = 1;
-
-    /** The primitive long 02. */
-    public final static long L_02 = 2;
-
-    /** The primitive long 03. */
-    public final static long L_03 = 3;
-
-    /** The primitive long 04. */
-    public final static long L_04 = 4;
-
-    /** The primitive long 05. */
-    public final static long L_05 = 5;
-
-    /** The primitive long 06. */
-    public final static long L_06 = 6;
-
-    /** The primitive long 07. */
-    public final static long L_07 = 7;
-
-    /** The primitive long 08. */
-    public final static long L_08 = 8;
-
-    /** The primitive long 09. */
-    public final static long L_09 = 9;
-
-    /** The primitive long 10. */
-    public final static long L_10 = 10;
-
-    /** The primitive long 100. */
-    public final static long L_100 = 100;
-
-    /** The primitive long 1000. */
-    public final static long L_1000 = 1000;
-
-    /** The primitive long 10000. */
-    public final static long L_10000 = 10000;
-
-    /** The primitive long 10001. */
-    public final static long L_10001 = 10001;
-
-    /** The primitive long 1001. */
-    public final static long L_1001 = 1001;
-
-    /** The primitive long 101. */
-    public final static long L_101 = 101;
-
-    /** The primitive long 11. */
-    public final static long L_11 = 11;
-
-    /** The primitive long 12. */
-    public final static long L_12 = 12;
-
-    /** The primitive long 13. */
-    public final static long L_13 = 13;
-
-    /** The primitive long 14. */
-    public final static long L_14 = 14;
-
-    /** The primitive long 15. */
-    public final static long L_15 = 15;
-
-    /** The primitive long 16. */
-    public final static long L_16 = 16;
-
-    /** The primitive long 17. */
-    public final static long L_17 = 17;
-
-    /** The primitive long 18. */
-    public final static long L_18 = 18;
-
-    /** The primitive long 19. */
-    public final static long L_19 = 19;
-
-    /** The primitive long 20. */
-    public final static long L_20 = 20;
-
-    /** The primitive long 21. */
-    public final static long L_21 = 21;
-
-    /** The primitive long 22. */
-    public final static long L_22 = 22;
-
-    /** The primitive long 23. */
-    public final static long L_23 = 23;
-
-    /** The primitive long 24. */
-    public final static long L_24 = 24;
-
-    /** The primitive long 25. */
-    public final static long L_25 = 25;
-
-    /** The primitive long 26. */
-    public final static long L_26 = 26;
-
-    /** The primitive long 27. */
-    public final static long L_27 = 27;
-
-    /** The primitive long 28. */
-    public final static long L_28 = 28;
-
-    /** The primitive long 29. */
-    public final static long L_29 = 29;
-
-    /** The primitive long 30. */
-    public final static long L_30 = 30;
-
-    /** The primitive long 31. */
-    public final static long L_31 = 31;
-
-    /** The primitive long 32. */
-    public final static long L_32 = 32;
-
-    /** The primitive long 33. */
-    public final static long L_33 = 33;
-
-    /** The primitive long 34. */
-    public final static long L_34 = 34;
-
-    /** The primitive long 35. */
-    public final static long L_35 = 35;
-
-    /** The primitive long 36. */
-    public final static long L_36 = 36;
-
-    /** The primitive long 37. */
-    public final static long L_37 = 37;
-
-    /** The primitive long 38. */
-    public final static long L_38 = 38;
-
-    /** The primitive long 39. */
-    public final static long L_39 = 39;
-
-    /** The primitive long 40. */
-    public final static long L_40 = 40;
-
-    /** The primitive long 41. */
-    public final static long L_41 = 41;
-
-    /** The primitive long 42. */
-    public final static long L_42 = 42;
-
-    /** The primitive long 43. */
-    public final static long L_43 = 43;
-
-    /** The primitive long 44. */
-    public final static long L_44 = 44;
-
-    /** The primitive long 45. */
-    public final static long L_45 = 45;
-
-    /** The primitive long 46. */
-    public final static long L_46 = 46;
-
-    /** The primitive long 47. */
-    public final static long L_47 = 47;
-
-    /** The primitive long 48. */
-    public final static long L_48 = 48;
-
-    /** The primitive long 49. */
-    public final static long L_49 = 49;
-
-    /** The primitive long 50. */
-    public final static long L_50 = 50;
-
-    /** The primitive long 51. */
-    public final static long L_51 = 51;
-
-    /** The primitive long 52. */
-    public final static long L_52 = 52;
-
-    /** The primitive long 53. */
-    public final static long L_53 = 53;
-
-    /** The primitive long 54. */
-    public final static long L_54 = 54;
-
-    /** The primitive long 55. */
-    public final static long L_55 = 55;
-
-    /** The primitive long 56. */
-    public final static long L_56 = 56;
-
-    /** The primitive long 57. */
-    public final static long L_57 = 57;
-
-    /** The primitive long 58. */
-    public final static long L_58 = 58;
-
-    /** The primitive long 59. */
-    public final static long L_59 = 59;
-
-    /** The primitive long 60. */
-    public final static long L_60 = 60;
-
-    /** The primitive long 61. */
-    public final static long L_61 = 61;
-
-    /** The primitive long 62. */
-    public final static long L_62 = 62;
-
-    /** The primitive long 63. */
-    public final static long L_63 = 63;
-
-    /** The primitive long 64. */
-    public final static long L_64 = 64;
-
-    /** The primitive long 65. */
-    public final static long L_65 = 65;
-
-    /** The primitive long 66. */
-    public final static long L_66 = 66;
-
-    /** The primitive long 67. */
-    public final static long L_67 = 67;
-
-    /** The primitive long 68. */
-    public final static long L_68 = 68;
-
-    /** The primitive long 69. */
-    public final static long L_69 = 69;
-
-    /** The primitive long 70. */
-    public final static long L_70 = 70;
-
-    /** The primitive long 71. */
-    public final static long L_71 = 71;
-
-    /** The primitive long 72. */
-    public final static long L_72 = 72;
-
-    /** The primitive long 73. */
-    public final static long L_73 = 73;
-
-    /** The primitive long 74. */
-    public final static long L_74 = 74;
-
-    /** The primitive long 75. */
-    public final static long L_75 = 75;
-
-    /** The primitive long 76. */
-    public final static long L_76 = 76;
-
-    /** The primitive long 77. */
-    public final static long L_77 = 77;
-
-    /** The primitive long 78. */
-    public final static long L_78 = 78;
-
-    /** The primitive long 79. */
-    public final static long L_79 = 79;
-
-    /** The primitive long 80. */
-    public final static long L_80 = 80;
-
-    /** The primitive long 81. */
-    public final static long L_81 = 81;
-
-    /** The primitive long 82. */
-    public final static long L_82 = 82;
-
-    /** The primitive long 83. */
-    public final static long L_83 = 83;
-
-    /** The primitive long 84. */
-    public final static long L_84 = 84;
-
-    /** The primitive long 85. */
-    public final static long L_85 = 85;
-
-    /** The primitive long 86. */
-    public final static long L_86 = 86;
-
-    /** The primitive long 87. */
-    public final static long L_87 = 87;
-
-    /** The primitive long 88. */
-    public final static long L_88 = 88;
-
-    /** The primitive long 89. */
-    public final static long L_89 = 89;
-
-    /** The primitive long 90. */
-    public final static long L_90 = 90;
-
-    /** The primitive long 91. */
-    public final static long L_91 = 91;
-
-    /** The primitive long 92. */
-    public final static long L_92 = 92;
-
-    /** The primitive long 93. */
-    public final static long L_93 = 93;
-
-    /** The primitive long 94. */
-    public final static long L_94 = 94;
-
-    /** The primitive long 95. */
-    public final static long L_95 = 95;
-
-    /** The primitive long 96. */
-    public final static long L_96 = 96;
-
-    /** The primitive long 97. */
-    public final static long L_97 = 97;
-
-    /** The primitive long 98. */
-    public final static long L_98 = 98;
-
-    /** The primitive long 99. */
-    public final static long L_99 = 99;
-
-    /** The primitive long 999. */
-    public final static long L_999 = 999;
-
-    /** The primitive long 9999. */
-    public final static long L_9999 = 9999;
-
-    /** A negative primitive long. */
-    public final static long L_NEGATIVE = -Math.abs(randomLong());
-
-    /** A positive primitive long. */
-    public final static long L_POSITIVE = Math.abs(randomLong());
-
-    /** A random primitive long. */
-    public final static long L_RANDOM = randomLong();
-
-    /** A small primitive long between 0-100. */
-    public final static long L_SMALL = randomBetween(0, 100);
+    public final static float F_SMALL = RandomIntegerUtil.randomBetween(0, 10000) / 100F;
 
     /** The primitive short 00. */
     public final static short S_00 = 0;
@@ -2006,5 +1335,5 @@ public class TestPrimitive
     public final static short S_RANDOM = randomShort();
 
     /** A small primitive short between 0-100. */
-    public final static short S_SMALL = (short) randomBetween(0, 100);
+    public final static short S_SMALL = (short) RandomIntegerUtil.randomBetween(0, 100);
 }
