+++ src/main/java/com/threeleaf/test/random/util/AbstractTestUtil.java
@@ -6,10 +6,8 @@ import lombok.Getter;
 import java.lang.reflect.Array;
 import java.util.*;
 
-import static com.threeleaf.test.random.TestPrimitive.*;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.*;
 import static com.threeleaf.test.random.util.TestBooleanUtil.randomBoolean;
-import static com.threeleaf.test.random.util.TestNumberUtil.random1to10;
-import static com.threeleaf.test.random.util.TestNumberUtil.randomBetween;
 import static com.threeleaf.test.random.util.TestSetUtil.convertToTreeSet;
 
 /**
@@ -39,6 +37,17 @@ public abstract class AbstractTestUtil<T>
         type = theType;
     }
 
+    /**
+     * Return a randomized object.
+     * This method may be overridden to provide customized randomization suitable
+     * to the application's test needs.
+     *
+     * @return a randomized object
+     */
+    public T random()
+    {
+        return TestRandom.random(type);
+    }
     /**
      * Return a randomized object.
      * This method may be overridden to provide customized randomization suitable
@@ -314,7 +323,7 @@ public abstract class AbstractTestUtil<T>
     public List<T> randomList(final int size, final String... fieldsExcluded)
     {
         List<T> list;
-        switch (randomBetween(INT_01, INT_03))
+        switch (RandomIntegerUtil.randomBetween(INT_01, INT_03))
         {
             case 1:
                 list = randomArrayList(size, fieldsExcluded);
@@ -365,7 +374,7 @@ public abstract class AbstractTestUtil<T>
     public Set<T> randomSet(final int size, final String... fieldsExcluded)
     {
         Set<T> set;
-        switch (randomBetween(INT_01, INT_03))
+        switch (RandomIntegerUtil.randomBetween(INT_01, INT_03))
         {
             case 1:
                 set = randomHashSet(size, fieldsExcluded);
