+++ src/main/java/com/threeleaf/test/random/util/TestAddressUtil.java
@@ -8,13 +8,11 @@ import lombok.experimental.UtilityClass;
 import java.util.List;
 import java.util.Map;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
-import static com.threeleaf.test.random.TestPrimitive.INT_10000;
 import static com.threeleaf.test.random.TestString.EMPTY;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.*;
 import static com.threeleaf.test.random.util.TestCollectionUtil.chooseOneFrom;
 import static com.threeleaf.test.random.util.TestMapUtil.chooseOneKeyFrom;
 import static com.threeleaf.test.random.util.TestMapUtil.chooseOneValueFrom;
-import static com.threeleaf.test.random.util.TestNumberUtil.randomBetween;
 
 /** Utilities for random address creation. */
 @UtilityClass
