+++ src/main/java/com/threeleaf/test/random/util/TestArrayUtil.java
@@ -5,10 +5,8 @@ import lombok.experimental.UtilityClass;
 
 import java.lang.reflect.Array;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_00;
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.*;
 import static com.threeleaf.test.random.util.TestListUtil.randomArrayListOf;
-import static com.threeleaf.test.random.util.TestNumberUtil.random1to10;
 import static java.util.Arrays.asList;
 
 /** Utilities to generate randomly populated Array objects. */
