+++ src/main/java/com/threeleaf/test/random/util/TestMapUtil.java
@@ -5,12 +5,9 @@ import lombok.experimental.UtilityClass;
 
 import java.util.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
-import static com.threeleaf.test.random.TestPrimitive.INT_03;
 import static com.threeleaf.test.random.TestRandom.random;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.*;
 import static com.threeleaf.test.random.util.TestCollectionUtil.chooseOneFrom;
-import static com.threeleaf.test.random.util.TestNumberUtil.random1to10;
-import static com.threeleaf.test.random.util.TestNumberUtil.randomBetween;
 import static com.threeleaf.test.random.util.TestSetUtil.randomHashSetOf;
 import static java.util.Comparator.comparing;
