+++ src/main/java/com/threeleaf/test/random/util/TestNumberUtil.java
@@ -5,89 +5,15 @@ import lombok.experimental.UtilityClass;
 import java.math.BigDecimal;
 import java.math.BigInteger;
 
-import static com.threeleaf.test.random.TestNumber.MAX_DIGIT;
-import static com.threeleaf.test.random.TestNumber.MIN_DIGIT;
-import static com.threeleaf.test.random.TestPrimitive.*;
 import static com.threeleaf.test.random.TestRandom.RANDOM;
-import static java.lang.Math.*;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_100;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_10000;
 
 /** Test Number Utilities. */
 @UtilityClass
 @SuppressWarnings("WeakerAccess")
 public class TestNumberUtil
 {
-
-    /**
-     * Get a random digit between 0 and 10.
-     *
-     * @return a number between 0 to 10 inclusive
-     */
-    public static int random0to10()
-    {
-        return randomBetween(0, INT_10);
-    }
-
-    /**
-     * Get a random number between 0 and 100.
-     *
-     * @return a number between 0 to 100 inclusive
-     */
-    public static int random0to100()
-    {
-        return randomBetween(0, INT_100);
-    }
-
-    /**
-     * Get a random digit between 1 and 10.
-     *
-     * @return a number between 1 to 10 inclusive
-     */
-    public static int random1to10()
-    {
-        return randomBetween(INT_01, INT_10);
-    }
-
-    /**
-     * Get a random number between 1 and 100.
-     *
-     * @return a number between 1 to 100 inclusive
-     */
-    public static int random1to100()
-    {
-        return randomBetween(INT_01, INT_100);
-    }
-
-    /**
-     * Return a random integer between two integers, inclusive.
-     *
-     * @param number1 a number
-     * @param number2 another number
-     *
-     * @return the random integer
-     */
-    public static int randomBetween(final int number1, final int number2)
-    {
-        final int min = min(number1, number2);
-        final int max = max(number1, number2);
-
-        return min + RANDOM.nextInt(max - min + 1);
-    }
-
-    /**
-     * Return a random long between two longs, inclusive.
-     *
-     * @param number1 a number
-     * @param number2 another number
-     *
-     * @return the random long
-     */
-    public static long randomBetween(final long number1, final long number2)
-    {
-        final long min = min(number1, number2);
-        final long max = max(number1, number2);
-
-        return min + (long) ((max - min + 1) * random());
-    }
 
     /**
      * Return a random {@link BigDecimal}.
@@ -116,17 +42,7 @@ public class TestNumberUtil
      */
     public static byte randomByte()
     {
-        return (byte) randomBetween(Byte.MIN_VALUE, Byte.MAX_VALUE);
-    }
-
-    /**
-     * Get a random digit.
-     *
-     * @return a number between 0 to 9 inclusive
-     */
-    public static int randomDigit()
-    {
-        return randomBetween(MIN_DIGIT, MAX_DIGIT);
+        return (byte) RandomIntegerUtil.randomBetween(Byte.MIN_VALUE, Byte.MAX_VALUE);
     }
 
     /**
@@ -150,33 +66,13 @@ public class TestNumberUtil
     }
 
     /**
-     * Return a random int.
-     *
-     * @return the int
-     */
-    public static int randomInt()
-    {
-        return RANDOM.nextInt();
-    }
-
-    /**
-     * Return a random int.
-     *
-     * @return the int
-     */
-    public static long randomLong()
-    {
-        return RANDOM.nextLong();
-    }
-
-    /**
      * Get a random number between 0.00 and 100.00.
      *
      * @return a number between 0.00 and 100.00 inclusive
      */
     public static double randomPercent()
     {
-        return (double) randomBetween(0, INT_10000) / (double) INT_100;
+        return (double) RandomIntegerUtil.randomBetween(0, INT_10000) / (double) INT_100;
     }
 
     /**
@@ -186,6 +82,6 @@ public class TestNumberUtil
      */
     public static short randomShort()
     {
-        return (short) randomBetween(Short.MIN_VALUE, Short.MAX_VALUE);
+        return (short) RandomIntegerUtil.randomBetween(Short.MIN_VALUE, Short.MAX_VALUE);
     }
 }
