+++ src/test/java/com/threeleaf/test/random/TestCollectionTest.java
@@ -5,8 +5,8 @@ import org.junit.Test;
 import java.lang.reflect.*;
 
 import static com.threeleaf.test.random.TestCollection.*;
-import static com.threeleaf.test.random.TestPrimitive.INT_00;
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_00;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_01;
 import static org.junit.Assert.*;
 
 /** Test {@link TestCollection}. */
