+++ src/test/java/com/threeleaf/test/random/util/TestAddressUtilTest.java
@@ -5,8 +5,8 @@ import org.junit.Test;
 
 import java.lang.reflect.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_02;
-import static com.threeleaf.test.random.TestPrimitive.INT_05;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_02;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_05;
 import static org.junit.Assert.*;
 
 /** Test {@link TestAddressUtil}. */
