+++ src/test/java/com/threeleaf/test/random/util/TestCollectionUtilTest.java
@@ -5,9 +5,9 @@ import org.junit.Test;
 import java.lang.reflect.*;
 import java.util.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
-import static com.threeleaf.test.random.TestPrimitive.INT_10;
 import static com.threeleaf.test.random.TestString.*;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_01;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_10;
 import static com.threeleaf.test.random.util.TestBooleanUtilTest.LOOP_COUNT_MAX;
 import static java.util.Arrays.asList;
 import static org.junit.Assert.*;
