+++ src/test/java/com/threeleaf/test/random/util/TestListUtilTest.java
@@ -4,7 +4,7 @@ import org.junit.Test;
 
 import java.lang.reflect.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_01;
 import static org.junit.Assert.*;
 
 /** Test {@link TestListUtil}. */
