+++ src/test/java/com/threeleaf/test/random/util/TestLoremIpsumUtilTest.java
@@ -4,9 +4,9 @@ import org.junit.Test;
 
 import java.lang.reflect.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_10;
 import static com.threeleaf.test.random.TestString.SPACE;
-import static com.threeleaf.test.random.util.TestNumberUtil.random1to10;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_10;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.random1to10;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
