+++ src/test/java/com/threeleaf/test/random/util/TestMapUtilTest.java
@@ -5,8 +5,8 @@ import org.junit.Test;
 import java.lang.reflect.*;
 import java.util.*;
 
-import static com.threeleaf.test.random.TestPrimitive.INT_01;
-import static com.threeleaf.test.random.TestPrimitive.INT_10;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_01;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.INT_10;
 import static com.threeleaf.test.random.util.TestBooleanUtilTest.LOOP_COUNT_MAX;
 import static org.junit.Assert.*;
