+++ src/test/java/com/threeleaf/test/random/util/TestNumberUtilTest.java
@@ -4,9 +4,9 @@ import org.junit.Test;
 
 import java.lang.reflect.*;
 
-import static com.threeleaf.test.random.TestNumber.LONG_05;
-import static com.threeleaf.test.random.TestNumber.LONG_10;
-import static com.threeleaf.test.random.TestPrimitive.*;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.*;
+import static com.threeleaf.test.random.util.RandomLongUtil.LONG_05;
+import static com.threeleaf.test.random.util.RandomLongUtil.LONG_10;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 
@@ -31,61 +31,61 @@ public class TestNumberUtilTest
         }
     }
 
-    /** Test {@link TestNumberUtil#random0to10()}. */
+    /** Test {@link RandomIntegerUtil#random0to10()}. */
     @Test
     public void random0to10()
     {
-        final int number = TestNumberUtil.random0to10();
+        final int number = RandomIntegerUtil.random0to10();
 
         assertTrue(number >= INT_00);
         assertTrue(number <= INT_10);
     }
 
-    /** Test {@link TestNumberUtil#random0to100()}. */
+    /** Test {@link RandomIntegerUtil#random0to100()}. */
     @Test
     public void random0to100()
     {
-        final int number = TestNumberUtil.random0to100();
+        final int number = RandomIntegerUtil.random0to100();
 
         assertTrue(number >= INT_00);
         assertTrue(number <= INT_100);
     }
 
-    /** Test {@link TestNumberUtil#random1to10()}. */
+    /** Test {@link RandomIntegerUtil#random1to10()}. */
     @Test
     public void random1to10()
     {
-        final int number = TestNumberUtil.random1to10();
+        final int number = RandomIntegerUtil.random1to10();
 
         assertTrue(number >= INT_01);
         assertTrue(number <= INT_10);
     }
 
-    /** Test {@link TestNumberUtil#random1to100()}. */
+    /** Test {@link RandomIntegerUtil#random1to100()}. */
     @Test
     public void random1to100()
     {
-        final int number = TestNumberUtil.random1to100();
+        final int number = RandomIntegerUtil.random1to100();
 
         assertTrue(number >= INT_01);
         assertTrue(number <= INT_100);
     }
 
-    /** Test {@link TestNumberUtil#randomBetween(int, int)}. */
+    /** Test {@link RandomIntegerUtil#randomBetween(int, int)}. */
     @Test
     public void randomBetween()
     {
-        final int number = TestNumberUtil.randomBetween(INT_05, INT_10);
+        final int number = RandomIntegerUtil.randomBetween(INT_05, INT_10);
 
         assertTrue(number >= INT_05);
         assertTrue(number <= INT_10);
     }
 
-    /** Test {@link TestNumberUtil#randomBetween(long, long)}. */
+    /** Test {@link RandomLongUtil#randomBetween(long, long)}. */
     @Test
     public void randomBetweenLong()
     {
-        final long number = TestNumberUtil.randomBetween(LONG_10, LONG_05);
+        final long number = RandomLongUtil.randomBetween(LONG_10, LONG_05);
 
         assertTrue(number >= LONG_05);
         assertTrue(number <= LONG_10);
@@ -98,11 +98,11 @@ public class TestNumberUtilTest
         assertNotNull(TestNumberUtil.randomBigDecimal());
     }
 
-    /** Test {@link TestNumberUtil#randomDigit()}. */
+    /** Test {@link RandomIntegerUtil#randomDigit()}. */
     @Test
     public void randomDigit()
     {
-        final int number = TestNumberUtil.randomDigit();
+        final int number = RandomIntegerUtil.randomDigit();
 
         assertTrue(number >= INT_00);
         assertTrue(number <= INT_09);
