+++ src/test/java/com/threeleaf/test/random/util/TestStringUtilTest.java
@@ -5,7 +5,7 @@ import org.junit.Test;
 import java.lang.reflect.*;
 
 import static com.threeleaf.test.random.TestString.*;
-import static com.threeleaf.test.random.util.TestNumberUtil.random1to10;
+import static com.threeleaf.test.random.util.RandomIntegerUtil.random1to10;
 import static com.threeleaf.test.random.util.TestStringUtil.STRING_LENGTH_MAX;
 import static com.threeleaf.test.random.util.TestStringUtil.STRING_LENGTH_MIN;
 import static org.junit.Assert.*;
